refactor: extract shared media device/severity metas hook (UN-2897)
Extract the video-count/device-breakdown/severity-tags computation shared
between MediaListMetaRow (campaign) and the upcoming hub meta row into
useMediaDeviceAndSeverityMetas, preserving MediaListMetaRow's existing
behavior exactly.
